+++ HISTORY.rst
@@ -16,6 +16,7 @@ History
     * Added new test to check RESTService.collection_get method (rudaporto).
     * Move database fixtures to conftest and use Base sqlalchemy from briefy.common (rudaporto).
     * Fix: validate column filters before trying to access attr to create the filter (rudaporto).
+    * Improve base resource filters to be filter association proxy attributes using special comparator (rudaporto).
 
 
 2.0.6 (2017-07-27)